spam-sniffer-backend/requirements.txt
Flask==3.1.0
flask-cors==4.0.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
sumy==0.11.0
nltk==3.9.1
requests==2.28.2
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.50.0
pymongo==4.3.3
